step angle should be float

+++ esp32_arm/src/CNC/Actuator/stepper/mechanic.h
@@ -9,7 +9,7 @@ class ActuatorMechanicStepper: public ActuatorMechanicBase{
 
     protected:
         int _micro_steps_on_stepper_driver;
-        int _motor_step_angle_in_degree = 1.8;
+        float _motor_step_angle_in_degree = 1.8;
         int _motor_gear_teeth_count;
         int _slave_pulley_teeth_count;
